pyproject.toml
[project]
name = "recipes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "markdown>=3.10.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pint>=0.25.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "termcolor>=3.3.0",
    "werkzeug>=3.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["recipe_core"]

[dependency-groups]
dev = [
    "snakeviz>=2.2.2",
]
